.history/requirements_20250522200727.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
httpx==0.25.1

# Machine Learning & Data Processing
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
monai==1.3.0
pydicom==2.4.3
opencv-python==4.8.1.78
scikit-image==0.22.0
scikit-learn==1.3.2
numpy==1.26.1
pandas==2.1.2

# NLP
transformers==4.35.0
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pytest==7.4.3 
